% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngkm_predict.R
\name{ngkm_predict}
\alias{ngkm_predict}
\title{ngkm_predict}
\usage{
ngkm_predict(input, model)
}
\arguments{
\item{input}{A string that will be used as predictor for the model.}

\item{model}{An ngk model.}
}
\value{
A data table with the predictions and their statistics.
}
\description{
This function returns predictions given an input text and a model. The number
of predictions will depend on the model
}
