% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngkm_prep.R
\name{ngkm_prep}
\alias{ngkm_prep}
\title{ngkm_prep}
\usage{
ngkm_prep(input, model = NULL, dictionary = NULL)
}
\arguments{
\item{input}{A string that will be tokenized.}

\item{model}{An ngk model. If an ngk model is provided, only the last n tokens
are going to be returned, being n the number of ngrams you used to train the
model decreased by 1.}
}
\value{
A list of tokens
}
\description{
This takes a raw string and tokenize it in a way it can be used to predict or
or to train an ngk model
}
