% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngkm_replace.R
\name{ngkm_replace}
\alias{ngkm_replace}
\title{ngkm_prep}
\usage{
ngkm_replace(
  input,
  what = c("url", "no", "_"),
  replacement = c("aurlthatwillbereplaced", "anothatwillbereplaced", " ")
)
}
\arguments{
\item{input}{A string that will be have their what(s) removed}

\item{what}{An "url", "no", "_" or any combination of these passed as a vector.}

\item{replacement}{A string or a list of strings that are going to be used to
replace the whats found in your text. If more than one, they should be listed
in the same order.}
}
\value{
A list of tokens
}
\description{
Just a useful tool to replace/remove "urls", "no" and "_" from the text. I
wish I had the stackoverflow treads to credit the kind people that wrote those
regular expressions but I lost them.
}
