% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngkm_train.R
\name{ngkm_train}
\alias{ngkm_train}
\title{ngkm_train}
\usage{
ngkm_train(
  token_object,
  ngrams = 2,
  lambda = 0.4,
  k = 0,
  npred = 3,
  no_pred_list = c("<unk>", "<no>", "<url>")
)
}
\arguments{
\item{token_object}{A token object from quanteda to be transformed into a dfm.}

\item{ngrams}{The maximum number of ngrams to used for training. It is
important to notice that if you train a model with ngrams = 5, the model will
be created with 4 predictors and 1 predicter tokens.}

\item{lambda}{The rate at which a prediction probability will be penalized for
having less than the maximum number of ngrams as predictors. The penalization
is exponential to the amount of missing matching tokens}

\item{k}{Any ngram that appears k or less times in the dataset is going to be
removed from the model}

\item{npred}{the number of prediction the model is going to return}

\item{no_pred_list}{a vector containing words that might be used as a predictor,
but will never be predicted by the model}
}
\value{
A list of tokens
}
\description{
Trains an ngk model.
Warning: can be very intense on your computer depending on the numbers you select.
}
