% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngkm_train.R
\docType{data}
\name{ngkmodel}
\alias{ngkmodel}
\title{A model trained using about 3.2 million documents}
\format{
a ngkm model:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{https://github.com/Harnkal/DS_CapstoneProject}
}
\usage{
ngkmodel
}
\description{
The model was trained with ngrams = 5, lambda = 0.75, k = 1 an npred = 6.
}
\keyword{datasets}
