% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngkm_train.R
\docType{data}
\name{ngkmodel}
\alias{ngkmodel}
\title{ngkmodel}
\format{
a ngkm model
}
\source{
\url{https://github.com/Harnkal/DS_CapstoneProject}
}
\usage{
ngkmodel
}
\description{
A model trained using about 3.2 million documents. The model was trained with
ngrams = 5, lambda = 0.75, k = 3 an npred = 6.
}
\keyword{datasets}
